% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{anzsco}
\alias{anzsco}
\title{ANZSCO}
\format{
A \code{tibble} with 11 variables:
\describe{
\item{\code{anzsco1_code}}{Major ANZSCO group codes, e.g. "4"}
\item{\code{anzsco1}}{Major ANZSCO group names in title case, e.g. "Community and Personal Service Workers"}
\item{\code{anzsco1_f}}{Factor version of \code{anzsco1}.}
\item{\code{anzsco2_code}}{Sub-major ANZSCO group codes, e.g. "45"}
\item{\code{anzsco2}}{Sub-major ANZSCO group names in title case, e.g. "Sports and Personal Service Workers"}
\item{\code{anzsco2_f}}{Factor version of \code{anzsco2}.}
\item{\code{anzsco3_code}}{Minor ANZSCO group codes, e.g. "452"}
\item{\code{anzsco3}}{Minor ANZSCO group names in title case, e.g. "Sports and Fitness Workers"}
\item{\code{anzsco3_f}}{Factor version of \code{anzsco3}.}
\item{\code{anzsco4_code}}{Unit ANZSCO group codes, e.g. "4523"}
\item{\code{anzsco4}}{Unit ANZSCO group names in title case, e.g. "Sports Coaches, Instructors and Officials"}
\item{\code{anzsco4_f}}{Factor version of \code{anzsco4}.}
\item{\code{anzsco6_code}}{Occupation ANZSCO group codes, e.g. "452311"}
\item{\code{anzsco6}}{Occupation ANZSCO group names in title case, e.g. "Diving Instructor (Open Water)"}
\item{\code{anzsco6_f}}{Factor version of \code{anzsco5}.}
\item{\code{skill_level}}{Skill level required for occupation, determined by the ABS (1 is highest, 5 is lowest).
See \url{https://www.abs.gov.au/ausstats/abs@.nsf/Previousproducts/C4BECE1704987586CA257089001A9181 } for details.}
}
}
\usage{
anzsco
}
\description{
Wide table containing all levels of the Australian and New Zealand Standard
Classification of Occupations (ANZSCO), First Edition, Revision 1, 2009. Cat. 1220.0.
}
\keyword{datasets}
