% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{anzsic}
\alias{anzsic}
\title{ANZSIC}
\format{
A \code{tibble} with 12 variables:
\describe{
\item{\code{anzsic_division_code}}{ANZSCO division codes character, e.g. "A", "B"}
\item{\code{anzsic_division}}{ANZSCO division title, e.g. "Agriculture, Forestry and Fishing"}
\item{\code{anzsic_division_f}}{Factor version of \code{anzsic_division}}
\item{\code{anzsic_subdivision_code}}{ANZSCO subdivision codes integer, e.g. 1, 2}
\item{\code{anzsic_subdivision}}{ANZSCO subdivision title, e.g. "Agriculture"}
\item{\code{anzsic_subdivision_f}}{Factor version of \code{anzsic_subdivision}}
\item{\code{anzsic_group_code}}{ANZSCO group codes integer, e.g. 11, 12}
\item{\code{anzsic_group}}{ANZSCO group title, e.g. "Mushroom and Vegetable Growing"}
\item{\code{anzsic_group_f}}{Factor version of \code{anzsic_group}}
\item{\code{anzsic_class_code}}{ANZSCO class codes integer, e.g. 111, 112}
\item{\code{anzsic_class}}{ANZSCO class title, e.g. "Vegetable Growing (Under Cover)"}
\item{\code{anzsic_class_f}}{Factor version of \code{anzsic_class}}
}
}
\usage{
anzsic
}
\description{
Wide table containing all levels of the Australian and New Zealand Standard
Industrial Classification (ANZSIC), 2006 (Revision 1.0). Cat. 1292.0.
}
\keyword{datasets}
