% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{anzsic2006}
\alias{anzsic2006}
\title{ANZSIC 2006}
\format{
A \code{tibble} with 8 variables:
\describe{
\item{\code{anzsic_division_code}}{ANZSIC division codes character, e.g. "A", "B"}
\item{\code{anzsic_division}}{ANZSIC division title, e.g. "Agriculture, Forestry and Fishing"}
\item{\code{anzsic_subdivision_code}}{ANZSIC subdivision codes 2-digit character, e.g. 01, 02}
\item{\code{anzsic_subdivision}}{ANZSIC subdivision title, e.g. "Agriculture"}
\item{\code{anzsic_group_code}}{ANZSIC group codes 3-digit character, e.g. 011, 012}
\item{\code{anzsic_group}}{ANZSIC group title, e.g. "Mushroom and Vegetable Growing"}
\item{\code{anzsic_class_code}}{ANZSIC class codes 4-digit character, e.g. 0111, 0112}
\item{\code{anzsic_class}}{ANZSIC class title, e.g. "Vegetable Growing (Under Cover)"}
}
}
\source{
\url{https://www.abs.gov.au/statistics/classifications/australian-and-new-zealand-standard-industrial-classification-anzsic/2006-revision-2-0/numbering-system-and-titles/division-subdivision-group-and-class-codes-and-titles}
}
\usage{
anzsic2006
}
\description{
Wide table containing all levels of the Australian and New Zealand Standard
Industrial Classification (ANZSIC), 2006 (Revision 2.0). Cat. 1292.0.
}
\keyword{datasets}
