% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asc_core_competencies}
\alias{asc_core_competencies}
\title{Australian Skills Classification (ASC) Core Competencies}
\format{
A \code{tibble} with:
\describe{
\item{\code{anzsco_code}}{Character. An ANZSCO unit or occupation code}
\item{\code{anzsco_name}}{Character. An ANZSCO unit or occupation name}
\item{\code{core_competencies}}{Character. One of "Numeracy", "Digital engagement", "Teamwork", "Writing", "Reading", "Learning", "Problem solving", "Oral communication", "Planning and organising", "Initiative and innovation". See description}
\item{\code{score}}{Integer. See description}
\item{\code{proficiency_level}}{Character. See description}
\item{\code{anchor_value}}{Character. See description}
}
}
\usage{
asc_core_competencies
}
\description{
As described by the National Skills Commission:
Core competencies are common to all jobs. They describe a set of non-specialist
skills gained in early life and schooling and provide a base to further
develop skills and specialties . Popular terms for these include ‘foundation
skills’, ‘common skills’, ‘soft skills’, ‘core skills’ and ‘employability
skills’. As part of the Australian Skills Classification work, the NSC
identified 10 core competencies required for every occupation in Australia.
The core competencies align to the definitions of foundation skills typically
used in the Australian VET system – specifically, the Employability Skills
Framework developed by the Australian Skills Quality Authority, with minor
differences recommended by education system experts.
}
\details{
Note: this data includes content of the Australian Skills Classification by
the National Skills Commission, Commonwealth of Australia.
Used under CC BY 4.0 licence.
}
\keyword{datasets}
