% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asc_descriptions}
\alias{asc_descriptions}
\title{Descriptions of ANZSCO used in the Australian Skills Classification (ASC)}
\format{
A \code{tibble} with:
\describe{
\item{\code{anzsco_code}}{Character. An ANZSCO unit or occupation code}
\item{\code{anzsco_name}}{Character. An ANZSCO unit or occupation name}
\item{\code{anzsco_description}}{Character. A description of the ANZSCO unit or occupation, provided by the ABS}
}
}
\usage{
asc_descriptions
}
\description{
As described by the National Skills Commission: ANZSCO Occupation codes and descriptions.
}
\details{
Note: this data includes content of the Australian Skills Classification by the National Skills Commission, Commonwealth of Australia. Used under CC BY 4.0 licence.
}
\keyword{datasets}
