% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asc_specialist_tasks}
\alias{asc_specialist_tasks}
\title{Australian Skills Classification (ASC) Specialist Tasks}
\format{
A \code{tibble} with:
\describe{
\item{\code{anzsco_code}}{Character. An ANZSCO unit or occupation code}
\item{\code{anzsco_name}}{Character. An ANZSCO unit or occupation name}
\item{\code{specialist_task}}{Character. See description}
\item{\code{percent_of_time_spent_on_task}}{Numeric. See description}
\item{\code{specialist_cluster}}{Character. See description}
\item{\code{percent_of_time_spent_on_cluster}}{Numeric. See description}
\item{\code{cluster_family}}{Character. See description}
\item{\code{percent_of_time_spent_on_family}}{Numeric. See description}
}
}
\usage{
asc_specialist_tasks
}
\description{
As described by the National Skills Commission:
The specialist tasks are designed to describe day- to -day work within an
occupation, and are work activities a person undertakes specific to a job.
While these skills can be transferrable across occupations and sectors,
unlike core competencies, they are not universal. The Australian Skills
Classification identifies specialist tasks for occupations in the Australian
labour market in a way that allows these tasks to be compared across
different occupations.
}
\details{
The skills clusters are clusters of similar specialist tasks. These tasks are
considered to be broadly transferable – if you can do one task in the
cluster, you can do the others. They show how skills are related and
connected to one another, and illustrates the transferability of skills
across occupations.

Note: this data includes content of the Australian Skills Classification by
the National Skills Commission, Commonwealth of Australia.
Used under CC BY 4.0 licence.
}
\keyword{datasets}
