% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asc_technology_tools}
\alias{asc_technology_tools}
\title{Australian Skills Classification (ASC) Technology Tools}
\format{
A \code{tibble} with:
\describe{
\item{\code{anzsco_code}}{Character. An ANZSCO unit or occupation code}
\item{\code{anzsco_name}}{Character. An ANZSCO unit or occupation name}
\item{\code{technology_tool}}{Character. See description}
}
}
\usage{
asc_technology_tools
}
\description{
As described by the National Skills Commission:
Technology tools are a technology, such as software or hardware, that is used
within an occupation. Common technology tools, (such as search engines and
email,) are featured across most occupations, and these are captured in the
core competency of digital engagement, rather than under this section. The
remaining technology tools are highly specialised and occupation-specific.
}
\details{
Note: this data includes content of the Australian Skills Classification by
the National Skills Commission, Commonwealth of Australia.
Used under CC BY 4.0 licence.
}
\keyword{datasets}
