% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asc_technology_tools_ranking}
\alias{asc_technology_tools_ranking}
\title{Australian Skills Classification (ASC) Technology Tools Ranking}
\format{
A \code{tibble} with:
\describe{
\item{\code{anzsco_code}}{Character. An ANZSCO unit or occupation code}
\item{\code{anzsco_name}}{Character. An ANZSCO unit or occupation name}
\item{\code{technology_tool_example}}{Character. See description}
\item{\code{technology_tool_ranking}}{Integer. See description}
}
}
\usage{
asc_technology_tools_ranking
}
\description{
For each technology tool, examples of that tool that are used in Australia
are provided.
}
\keyword{datasets}
