% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asced_qual}
\alias{asced_qual}
\title{ASCED Level of Education}
\format{
A \code{tibble} with 6 variables:
\describe{
\item{\code{qual1_code}}{Broad qualification level one-digit code, e.g. "1"}
\item{\code{qual1}}{Broad qualification level one-digit name in title case, e.g. "Postgraduate Degree Level"}
\item{\code{qual2_code}}{Narrow qualification level two-digit code, e.g. "12"}
\item{\code{qual2}}{Narrow qualification level two-digit name in title case, e.g. "Master Degree Level"}
\item{\code{qual3_code}}{Detailed qualification level three-digit code, e.g. "122"}
\item{\code{qual3}}{Detailed qualification level three-digit name in title case, e.g. "Master Degree by Coursework"}
}
}
\usage{
asced_qual
}
\description{
Wide table containing all levels of qualifications in the Australian Standard
Classification of Education (ASCED), 2001. Cat. 1272.0.
}
\keyword{datasets}
