% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auholidays.R
\docType{data}
\name{auholidays}
\alias{auholidays}
\title{Australian Public Holidays}
\format{
A data frame with 3 variables:
\describe{
  \item{\code{Date}}{The date of the holiday}
  \item{\code{Name}}{The name of the holiday}
  \item{\code{Jurisdiction}}{The jurisdiction (state/territory code) to which the holiday applies}
}
}
\source{
Historical data (2014-2025) from \url{https://data.gov.au/data/dataset/australian-holidays-machine-readable-dataset}.
2026+ data compiled from state/territory government websites.
}
\usage{
auholidays
}
\description{
Tidy data set of Australian Public Holidays
}
\keyword{datasets}
