% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_seasons.R
\name{aus_seasons}
\alias{aus_seasons}
\title{Get Australian seasons for given dates}
\usage{
aus_seasons(dates, type = "temperate")
}
\arguments{
\item{dates}{vectors with dates. It supports Dates, POSIxt  or characters.}

\item{type}{either "temperate" (default) or "tropical"}
}
\value{
vector with season names
}
\description{
Small helper function that provided a vector with dates, returns the names of corresponding seasons, according to the Australian definition of seasons
It can return either "temperate" (4 seasons) or "tropical" (dry/wet) seasons.
}
\examples{
\dontrun{
sample_dates <- c("2022-01-01","2018-10-04","2016-07-20")
# temperate seasons
aus_seasons(sample_dates,"temperate")
# tropical seasons
aus_seasons(sample_dates,"tropical")
}
}
