% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_anzsco.R
\name{clean_anzsco}
\alias{clean_anzsco}
\title{Convert ANZSCO names and abbreviations into the consistent format used by the Australian Bureau of Statistics.
This function enables both exact (default) and fuzzy matching.
Under exact matching, if no match is found, NA is returned.}
\usage{
clean_anzsco(
  x,
  fuzzy_match = FALSE,
  max_dist = 0.4,
  method = "jw",
  silent = FALSE
)
}
\arguments{
\item{x}{a (character) vector containing ANZSCO titles.
Note that \code{clean_anzsco} always returns a character vector. If no match is found, then \code{NA} is returned.}

\item{fuzzy_match}{logical; either TRUE which indicates that
approximate/fuzzy string matching should be used, or FALSE (the default) which indicates that
only exact matches should be used. If FALSE, then if no match is found, then NA is returned.}

\item{max_dist}{numeric, sets the maximum acceptable distance between your
string and the matched string. Default is 0.4. Only relevant when fuzzy_match is TRUE.}

\item{method}{the method used for approximate/fuzzy string matching. Default
is "jw", the Jaro-Winker distance; see `??stringdist-metrics` for more options.
Only relevant when fuzzy_match is TRUE.}

\item{silent}{a boolean value. If FALSE (the default), the function will warn that  \code{NA}(s) were returned.}
}
\value{
a character vector of cleaned values that have matches in the official format.
}
\description{
Convert ANZSCO names and abbreviations into the consistent format used by the Australian Bureau of Statistics.
This function enables both exact (default) and fuzzy matching.
Under exact matching, if no match is found, NA is returned.
}
\examples{

clean_anzsco("driving instructor")

clean_anzsco("Driving", fuzzy_match = TRUE)

}
\seealso{
\code{\link{clean_anzsic}} for ANZSIC.

Other cleaning functions: 
\code{\link{clean_anzsic}()},
\code{\link{clean_asced_foe}()},
\code{\link{clean_asced_qual}()}
}
\concept{cleaning functions}
