% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nfd.R
\name{clean_nfd}
\alias{clean_nfd}
\title{Return highest level of detail from "NFD" codes}
\usage{
clean_nfd(vector)
}
\arguments{
\item{vector}{A character vector or column containing ANZCO/ANZSIC names or codes.}
}
\value{
Returns the same character vector without "NFD" or trailing zeroes. Strings without "NFC" or trailing zeroes are unchanged.
}
\description{
Removes ", nfd" or trailing zeroes from the end of ANZCO/ANZSIC labels or codes.
}
\examples{

clean_nfd("Furniture Manufacturing, nfd")

clean_nfd(c("Managers, nfd", "2510"))

}
\seealso{
Other cleaning functions: 
\code{\link{clean_anzsco}()},
\code{\link{clean_anzsic}()},
\code{\link{clean_asced_foe}()},
\code{\link{clean_asced_qual}()}
}
\concept{cleaning functions}
