% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_helpers.R
\name{clean_string}
\alias{clean_string}
\title{Helper function to clean strings}
\usage{
clean_string(string)
}
\arguments{
\item{string}{a character vector}
}
\value{
a cleaned string with no punctuation, trailing letter s, numbers etc
}
\description{
Helper function to clean strings
}
