% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\name{get_seifa}
\alias{get_seifa}
\title{Import SEIFA Data from ABS}
\usage{
get_seifa(
  structure = c("sa1", "sa2", "lga", "postcode", "suburb"),
  data_subclass = c("irsed", "irsead", "ier", "ieo"),
  year = NULL,
  method = Sys.getenv("R_STRAYR_DL_METHOD", unset = "auto")
)
}
\arguments{
\item{structure}{character value for the desired spatial area. Must be one of:
\describe{
  \item{sa1}{ - download size 51.6 MB}
  \item{sa2}{ - download size 1.9 MB}
  \item{lga}{ - download size 660 KB}
  \item{postcode}{ - download size 2.3 MB}
  \item{suburb}{ - download size 11.3 MB}
}}

\item{data_subclass}{character vector matching available SEIFA indexes:
\describe{
  \item{irsed}{ - Index of Relative Socio-economic Disadvantage}
  \item{irsead}{ - Index of Relative Socio-economic Advantage and Disadvantage}
  \item{ier}{ - Index of Economic Resources}
  \item{ieo}{ - Index of Education and Occupation}
}}

\item{year}{a character string or numeric of the release year of SEIFA object, eg "2016"; 2011.}

\item{method}{character, default taken from
\code{Sys.getenv("R_STRAYR_DL_METHOD", unset = "auto")}. The value is
passed to \code{download.file(method = ...)} when retrieving files from
GitHub. This is provided to help users on corporate networks where only a
specific download method is permitted by the proxy. For example on Windows,
\code{method = "wininet"} may be required.}
}
\value{
`data.frame` if file successfully downloaded, else returns `NULL`.
}
\description{
The function will download all SEIFA data, for a specified spatial structure,
to a temporary excel file and then merge sheets into a single `data.frame`. This `data.frame`
also includes the ABS population count for the given spatial structure. For more information
on SEIFA indexes go to
\url{https://www.abs.gov.au/statistics/people/people-and-communities/socio-economic-indexes-areas-seifa-australia}

To learn more about the individual data-sets for each year, please visit their respective pages:
2011
\url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/allprimarymainfeatures/8C5F5BB699A0921CCA258259000BA619}
2016
\url{https://www.abs.gov.au/ausstats/abs@.nsf/mf/2033.0.55.001}
2021
\url{https://www.abs.gov.au/statistics/people/people-and-communities/socio-economic-indexes-areas-seifa-australia/2021}
}
\details{
Some corporate networks restrict downloads from or require a specific
  HTTP stack. If \code{read_absmap()} times out when
  validating or downloading, set the environment variable
  \code{R_STRAYR_DL_METHOD} before calling this function, for example:
  \preformatted{Sys.setenv(R_STRAYR_DL_METHOD = "wininet")}
  You can add \code{R_STRAYR_DL_METHOD = "wininet"} to your \code{.Renviron}
  to persist across sessions.
}
\note{
For All ABS SEIFA spreadsheets go to \href{https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2033.0.55.0012016?OpenDocument}{ABS website}
}
\examples{
\dontrun{
get_seifa(structure = "lga", data_subclass = "irsed", year = 2016)
}

}
