% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\name{get_seifa_index_sheet}
\alias{get_seifa_index_sheet}
\title{Parse SEIFA index Spreadsheet}
\usage{
get_seifa_index_sheet(
  filename,
  sheetname,
  spatial_area = c("sa1", "sa2", "sa3", "lga", "postcode")
)
}
\arguments{
\item{filename}{character name of temporary file when spreadsheet downloaded}

\item{sheetname}{character name of the sheet to be imported}

\item{spatial_area}{character}
}
\value{
data.frame
}
\description{
This is a helper function used by \code{\link{seifa_scores}}. The function can also
be used independently if you have already downloaded one of the SEIFA index score spreadsheets
from link{https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2033.0.55.0012016?OpenDocument}
}
