% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_holiday.R
\name{is_holiday}
\alias{is_holiday}
\title{is_holiday}
\usage{
is_holiday(date, jurisdictions = c())
}
\arguments{
\item{date}{Date, POSIXct object or a string that can be parsed with `parsedate::parse_date`}

\item{jurisdictions}{Vector of state jurisdictions you wish to filter on, defaults to all of Australia}
}
\value{
logical(`TRUE` or `FALSE`) vector
}
\description{
Identifies if a given date is a public holiday in Australia
}
\examples{
is_holiday('2020-01-01')
is_holiday('2019-05-27', jurisdictions=c('ACT', 'TAS'))
h_df <- data.frame(dates = c('2020-01-01', '2020-01-10'))
h_df \%>\%
  dplyr::mutate(IsHoliday = is_holiday(dates))
}
