% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_helpers.R
\name{make_dictionary}
\alias{make_dictionary}
\title{A helper function to create a dictionary for use in the clean family of functions}
\usage{
make_dictionary(category_tbl)
}
\arguments{
\item{category_tbl}{}
}
\value{
a named vector.
}
\description{
A helper function to create a dictionary for use in the clean family of functions
}
