% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{palette_state_name_2016}
\alias{palette_state_name_2016}
\title{palette_state_name_2016}
\format{
A named \code{vector} with hexadecimal colours.
}
\usage{
palette_state_name_2016
}
\description{
Named vector with the 'unofficial colours of each State and Territory in Australia.
}
\keyword{datasets}
