% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_income.R
\name{parse_income_range}
\alias{parse_income_range}
\title{parse_income_range extracts dollar figures from income bucket ranges used in Census
data and elsewhere}
\usage{
parse_income_range(
  income_range,
  limit = "lower",
  max_income = Inf,
  is_zero = c("Nil income"),
  negative_as_zero = TRUE,
  dollar_prefix = "$",
  .silent = TRUE,
  .print_lookup = FALSE
)
}
\arguments{
\item{income_range}{a character vector containing income ranges, e.g.
\code{"b. $18,201 to $37,000"}, \code{"$400-$499 ($20,800-$25,999)"}}

\item{limit}{either "lower" (the default) and "upper" bounds of the range, or "mid" to return the mid-point}

\item{max_income}{numeric value of 'and over' value of the highest bracket. \code{Inf} by default.
If \code{NULL}, returns a max_income value equal to the lower bound of the
highest bracket. Important for \code{limit = "mid"} or \code{limit = "upper"}.}

\item{is_zero}{a character vector of values to be interpreted as zero. Defaults to \code{c("Nil income")}.}

\item{negative_as_zero}{a boolean value. If TRUE (the default) the character}

\item{dollar_prefix}{a character string to denote the dollar delimiter.
\code{"$"} by default, but could be e.g. \code{"AUD"} or \code{"USD"}.
string "Negative income" is interpreted as \code{0}. If FALSE, \code{NA} is returned.}

\item{.silent}{a boolean value. If FALSE (the default), the function will warn that  \code{NA}(s) were returned.}

\item{.print}{If TRUE prints lookup table displaying inputted income range and the returned parsed value}
}
\value{
a numeric vector of incomes
}
\description{
parse_income_range extracts dollar figures from income bucket ranges used in Census
data and elsewhere
}
