% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_absmaps.R
\name{read_absmap}
\alias{read_absmap}
\title{Read ABS geographic data}
\arguments{
\item{name}{a character string containing absmapsdata file names in [\code{area}][\code{year}] format, eg "sa42016"; "gcc2021".
See full list at https://github.com/wfmackey/absmapsdata.
Note: if \code{name} is entered, then \code{area} and \code{year} values will be ignored.}

\item{area}{a character string containing the concise absmapsdata area names, eg "sa4"; "gcc".
See full list at https://github.com/wfmackey/absmapsdata.}

\item{year}{a character string or numeric of the full source year of absmapsdata object, eg "2016"; 2021.
See full list at https://github.com/wfmackey/absmapsdata.}

\item{remove_year_suffix}{logical defaulting to FALSE.
If TRUE, `strip_year_suffix` is run before returning the object, removing
the `_year` suffix from variable names.}

\item{export_dir}{path to a directory to store the desired sf object. \code{tempdir()} by default.}

\item{.validate_name}{logical defaulting to TRUE, which checks the name input (or area year combination) against
a list of available objects in the \code{absmapsdata} package.}

\item{method}{character, default taken from
\code{Sys.getenv("R_STRAYR_DL_METHOD", unset = "auto")}. The value is
passed to \code{download.file(method = ...)} when retrieving files from
GitHub. This is provided to help users on corporate networks where only a
specific download method is permitted by the proxy. For example on Windows,
\code{method = "wininet"} may be required.}
}
\value{
an sf object.
}
\description{
Read ABS geographic data as sf objects from the [absmapsdata](https://github.com/wfmackey/absmapsdata) package.
}
\details{
Some corporate networks restrict downloads from or require a specific
  HTTP stack. If \code{read_absmap()} times out when
  validating or downloading, set the environment variable
  \code{R_STRAYR_DL_METHOD} before calling this function, for example:
  \preformatted{Sys.setenv(R_STRAYR_DL_METHOD = "wininet")}
  You can add \code{R_STRAYR_DL_METHOD = "wininet"} to your \code{.Renviron}
  to persist across sessions.
}
\examples{

\dontrun{read_absmap("sa42016")}

\dontrun{read_absmap(area = "sa4", year = 2016)}

}
\seealso{
https://github.com/wfmackey/absmapsdata
}
