% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_correspondence_tbl.R
\name{read_correspondence_tbl}
\alias{read_correspondence_tbl}
\title{Read in ASGC/ASGS 2016 population-weighted correspondence tables from the ABS.}
\source{
<https://data.gov.au/dataset/ds-dga-23fe168c-09a7-42d2-a2f9-fd08fbd0a4ce>
}
\usage{
read_correspondence_tbl(
  from_area,
  from_year,
  to_area,
  to_year,
  export_dir = tempdir(),
  method = Sys.getenv("R_STRAYR_DL_METHOD", unset = "auto")
)
}
\arguments{
\item{from_area}{The area you want to correspond FROM (ie the areas your data are currently in). For example: "sa1", "sa2, "sa3", "sa4".}

\item{from_year}{The year you want to correspond FROM. For example: 2011, 2016.}

\item{to_area}{The area you want to correspond TO (ie the areas you want your data to be in).}

\item{to_year}{The year you want to correspond TO.}

\item{export_dir}{path to a directory to store the desired sf object. \code{tempdir()} by default.}

\item{method}{character, default taken from
\code{Sys.getenv("R_STRAYR_DL_METHOD", unset = "auto")}. The value is
passed to \code{download.file(method = ...)} when retrieving files from
GitHub. This is provided to help users on corporate networks where only a
specific download method is permitted by the proxy. For example on Windows,
\code{method = "wininet"} may be required.}
}
\value{
A \code{tibble} object.
}
\description{
Read in ASGC/ASGS 2016 population-weighted correspondence tables from the ABS.
}
\details{
Some corporate networks restrict downloads from or require a specific
  HTTP stack. If \code{read_absmap()} times out when
  validating or downloading, set the environment variable
  \code{R_STRAYR_DL_METHOD} before calling this function, for example:
  \preformatted{Sys.setenv(R_STRAYR_DL_METHOD = "wininet")}
  You can add \code{R_STRAYR_DL_METHOD = "wininet"} to your \code{.Renviron}
  to persist across sessions.
}
\examples{
\dontrun{
sa4_corr <- read_correspondence_tbl("sa4", 2011, "sa4", 2016)
lga2011_to_2018 <- read_correspondence_tbl("LGA", 2011, "LGA", 2018)
}
}
