% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\name{seifa_scores}
\alias{seifa_scores}
\title{Import SEIFA Data from ABS Spreadsheet}
\usage{
seifa_scores(
  indexes = c("irsed", "irsead", "ier", "ieo"),
  spatial_area = c("sa1", "sa2", "sa3", "lga", "postcode")
)
}
\arguments{
\item{indexes}{character vector matching available SEIFA indexes:
\itemize{
  \item{irsed}{Index of Relative Socio-economic Disadvantage}
  \item{irsead}{Index of Relative Socio-economic Advantage and Disadvantage}
  \item{ier}{Index of Economic Resources}
  \item{ieo}{Index of Education and Occupation}
}}

\item{spatial_area}{character value for the desired spatial area. MUst be one of:
\itemize{
  \item{sa1}{download size 51.6 mb}
  \item{sa2}{download size 1.9 mb}
  \item{lga}{download size 660 kb}
  \item{postcode}
  \item{suburb}
}}
}
\description{
The function will download all LGA SEIFA data to a temporary excel file and merge sheets into a long `data.frame`
}
\note{
For All SEIFA spreadsheets go to https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2033.0.55.0012016?OpenDocument
}
