% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{state_abb_au}
\alias{state_abb_au}
\title{state_abb_au}
\format{
A character vector of Australian state and territory abbreviations.
}
\usage{
state_abb_au
}
\description{
Vector of Australian State and Territory abbreviations The Australian
equivalent of `datasets::state.abb`. All capitalised, as in "VIC".
}
\keyword{datasets}
