% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{state_name_au}
\alias{state_name_au}
\title{state_name_au}
\format{
A character vector of Australian state and territory names.
}
\usage{
state_name_au
}
\description{
Vector of Australian State and Territory names. The Australian
equivalent of `datasets::state.name`
}
\keyword{datasets}
