% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{state_or_territory_palette}
\alias{state_or_territory_palette}
\title{state_or_territory_palette}
\format{
A named \code{vector} with hexadecimal colours.
}
\usage{
state_or_territory_palette
}
\description{
Named vector with the 'unofficial colours of each State and Territory in Australia.
}
\keyword{datasets}
