% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_year_suffix.R
\name{strip_year_suffix}
\alias{strip_year_suffix}
\title{strip_year_suffix removes year suffixes from variable names}
\usage{
strip_year_suffix(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a tibble
}
\description{
Some structures in strayr contain variables with year
suffixes (e.g. \code{sa3_name_2011}). This function removes these year
suffixes from the dataset supplied.
}
