% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{asced_foe}
\alias{asced_foe}
\title{ASCED Field of Education}
\format{
A \code{tibble} with 6 variables:
\describe{
\item{\code{aced_foe_broad_code}}{Broad field of education two-digit code, e.g. "09"}
\item{\code{aced_foe_broad}}{Broad field of education two-digit name in title case, e.g. "Society and Culture"}
\item{\code{aced_foe_narrow_code}}{Narrow field of education four-digit code, e.g. "0921"}
\item{\code{aced_foe_narrow}}{Narrow field of education four-digit name in title case, e.g. "Sport and Recreation"}
\item{\code{aced_foe_detailed_code}}{Detailed field of education six-digit code, e.g. "092103"}
\item{\code{aced_foe_detailed}}{Detailed field of education six-digit name in title case, e.g. "Sports Coaching, Officiating and Instruction"}
}
}
\usage{
asced_foe
}
\description{
Wide table containing all levels of fields of education in the Australian
Standard Classification of Education (ASCED), 2001. Cat. 1272.0.
}
\keyword{datasets}
