% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auholidays.R
\docType{data}
\name{auholidays}
\alias{auholidays}
\title{Australian Public Holidays}
\format{
A data frame with 776 rows and 3 variables:
\describe{
  \item{\code{Date}}{The date of the holiday}
  \item{\code{Name}}{The name of the holiday}
  \item{\code{Jurisdiction}}{Pipe separated list of jurisdictions to which the holiday applies}
}
}
\source{
\url{https://data.gov.au/data/dataset/australian-holidays-machine-readable-dataset}
}
\usage{
auholidays
}
\description{
Tidy data set of Australian Public Holidays
}
\keyword{datasets}
