% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_helpers.R
\name{clean_titles}
\alias{clean_titles}
\title{clean_titles}
\usage{
clean_titles(
  dictionary,
  .vector = vector,
  .fuzzy_match = fuzzy_match,
  .max_dist = max_dist,
  .method = method,
  .silent = silent
)
}
\arguments{
\item{.silent}{}
}
\value{

}
\description{
clean_titles
}
\examples{

}
