% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\name{get_seifa_index_sheet}
\alias{get_seifa_index_sheet}
\title{Parse SEIFA index Spreadsheet}
\usage{
get_seifa_index_sheet(
  filename,
  sheetname,
  structure = c("sa1", "sa2", "lga", "postcode", "suburb"),
  year
)
}
\arguments{
\item{filename}{character name of temporary file when spreadsheet downloaded}

\item{sheetname}{character name of the sheet to be imported}

\item{structure}{character spatial structure of the data to be parsed. The spatial structure is
important as the shape of the data in the ABS spreadsheets if different for some structures.}

\item{year}{a character string or numeric of the release year of SEIFA object, eg "2016"; 2011.}
}
\value{
data.frame
}
\description{
This is a helper function used by \code{\link{get_seifa}}. The function can also
be used independently if you have already downloaded one of the SEIFA index score spreadsheets
from \url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2033.0.55.0012016?OpenDocument}
}
\examples{
\dontrun{

  get_seifa_index_sheet('downloaded_filename.xls', sheetname = 'Table 2', structure = 'lga')
}

}
