% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_holiday.R
\name{is_holiday}
\alias{is_holiday}
\title{is_holiday}
\usage{
is_holiday(date, jurisdictions = c())
}
\arguments{
\item{date}{Date, POSIXct object or a string that can be parsed with `parsedate::parse_date`}

\item{jurisdictions}{Vector of state jurisdictions you wish to filter on, defaults to all of Australia}
}
\value{
Returns a TRUE or FALSE value
}
\description{
Identifies if a given date is a public holiday in Australia
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
