% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_absmaps.R
\name{read_absmap}
\alias{read_absmap}
\title{Read ABS geographic data}
\usage{
read_absmap(name = NULL, area = NULL, year = NULL)
}
\arguments{
\item{name}{a character string containing absmapsdata file names in [\code{area}][\code{year}] format, eg "sa42016"; "gcc2021".
See full list at https://github.com/wfmackey/absmapsdata.
Note: if \code{name} is entered, then \code{area} and \code{year} values will be ignored.}

\item{area}{a character string containing the concise absmapsdata area names, eg "sa4"; "gcc".
See full list at https://github.com/wfmackey/absmapsdata.}

\item{year}{a character string or numeric of the full source year of absmapsdata object, eg "2016"; 2021.
See full list at https://github.com/wfmackey/absmapsdata.}
}
\value{
an sf object.
}
\description{
Read ABS geographic data as sf objects from the [absmapsdata](https://github.com/wfmackey/absmapsdata) package.
}
\examples{

\dontrun{read_absmap("sa42016")}

\dontrun{read_absmap(area = "sa4", year = 2016)}

}
\seealso{
https://github.com/wfmackey/absmapsdata
}
