% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_absmaps.R
\name{read_absmap}
\alias{read_absmap}
\title{Read ABS geographic data}
\arguments{
\item{name}{a character string containing absmapsdata file names in [\code{area}][\code{year}] format, eg "sa42016"; "gcc2021".
See full list at https://github.com/wfmackey/absmapsdata.
Note: if \code{name} is entered, then \code{area} and \code{year} values will be ignored.}

\item{area}{a character string containing the concise absmapsdata area names, eg "sa4"; "gcc".
See full list at https://github.com/wfmackey/absmapsdata.}

\item{year}{a character string or numeric of the full source year of absmapsdata object, eg "2016"; 2021.
See full list at https://github.com/wfmackey/absmapsdata.}

\item{remove_year_suffix}{logical defaulting to FALSE.
If TRUE, `strip_year_suffix` is run before returning the object, removing
the `_year` suffix from variable names.}

\item{export_dir}{path to a directory to store the desired sf object. \code{tempdir()} by default.}

\item{.validate_name}{logical defaulting to TRUE, which checks the name input (or area year combination) against
a list of available objects in the \code{absmapsdata} package.}
}
\value{
an sf object.
}
\description{
Read ABS geographic data as sf objects from the [absmapsdata](https://github.com/wfmackey/absmapsdata) package.
}
\examples{

\dontrun{read_absmap("sa42016")}

\dontrun{read_absmap(area = "sa4", year = 2016)}

}
\seealso{
https://github.com/wfmackey/absmapsdata
}
