% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_correspondence_tbl.R
\name{read_correspondence_tbl}
\alias{read_correspondence_tbl}
\title{Read in ASGC/ASGS 2016 population-weighted correspondence tables from the ABS.}
\source{
<https://data.gov.au/dataset/ds-dga-23fe168c-09a7-42d2-a2f9-fd08fbd0a4ce>
}
\usage{
read_correspondence_tbl(
  from_area,
  from_year,
  to_area,
  to_year,
  export_dir = tempdir()
)
}
\arguments{
\item{from_area}{The area you want to correspond FROM (ie the areas your data are currently in). For example: "sa1", "sa2, "sa3", "sa4".}

\item{from_year}{The year you want to correspond FROM. For example: 2011, 2016.}

\item{to_area}{The area you want to correspond TO (ie the areas you want your data to be in).}

\item{to_year}{The year you want to correspond TO.}

\item{export_dir}{path to a directory to store the desired sf object. \code{tempdir()} by default.}
}
\value{
A \code{tibble} object.
}
\description{
Read in ASGC/ASGS 2016 population-weighted correspondence tables from the ABS.
}
\examples{
\dontrun{
sa4_corr <- read_correspondence_tbl("sa4", 2011, "sa4", 2016)
lga2011_to_2018 <- read_correspondence_tbl("LGA", 2011, "LGA", 2018)
}
}
