% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strayr.R
\name{strayr}
\alias{strayr}
\title{Convert Australian state names and abbreviations into a consistent format}
\usage{
strayr(x, to = "state_abbr", fuzzy_match = TRUE, max_dist = 0.4,
  method = "jw")
}
\arguments{
\item{x}{a (character) vector containing Australian state names or abbreviations or
a (numeric) vector containing state codes (1 = NSW, 2 = Vic, 3 = Qld, 4 = SA,
5 = WA, 6 = Tas, 7 = NT, 8 = ACT). Note that strayr always returns a character vector.}

\item{to}{what form should the state names be converted to? Options are
"state_name", "state_abbr" (the default), "iso", "postal", and "code".}

\item{fuzzy_match}{logical; either TRUE (the default) which indicates that
approximate/fuzzy string matching should be used, or FALSE which indicates that
only exact matches should be used.}

\item{max_dist}{numeric, sets the maximum acceptable distance between your
string and the matched string. Default is 0.4. Only relevant when fuzzy_match is TRUE.}

\item{method}{the method used for approximate/fuzzy string matching. Default
is "jw", the Jaro-Winker distance; see `??stringdist-metrics` for more options.}
}
\value{
a character vector
}
\description{
Convert Australian state names and abbreviations into a consistent format
}
\examples{

x <- c("western Straya", "w. A ", "new soth wailes", "SA", "tazz")

# Convert the above to state abbreviations
strayr(x)

# Convert the elements of `x` to state names

strayr(x, to = "state_name")

# Disable fuzzy matching; you'll get NAs unless exact matches can be found

strayr(x, fuzzy_match = FALSE)

# You can use strayr in a dplyr mutate call

x_df <- data.frame(state = x, stringsAsFactors = FALSE)

\dontrun{x_df \%>\% mutate(state_abbr = strayr(state))}

}
